Ability to get config.
This will make it able to get more config lines from the file,
for example I gonna make some handlers for MM wich are going to 
hook into my plug CybsEnergy, but it would also be nice if the
users can easy change the energy healed (for example) to an amount
they want without any java knowledge. And you don't have to make 
a handler for every diffirent amount.